% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_predict.R
\name{cw_predict}
\alias{cw_predict}
\title{Predict a biomarker based on the other using specified model}
\usage{
cw_predict(dat, studies, mod_obj, N = 2000, seed = 2022)
}
\arguments{
\item{dat}{The data frame after transformation}

\item{studies}{A column for all study IDs}

\item{mod_obj}{A model object}

\item{N}{Number of simulations}

\item{seed}{Seed value for reproducibility; use default if not provided}
}
\value{
A vector of predicted probabilities
}
\description{
Predict a biomarker based on the other using specified model
}
\examples{
predicted_a1c <- cw_predict(data, data$study_id, mod_selected, N = 2000)
}
