% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_predict.R
\name{predict_missing_fpg}
\alias{predict_missing_fpg}
\title{Predict missing FPG based on HbA1c using an appropriate model}
\usage{
predict_missing_fpg(
  d,
  all_model_coefs,
  N_sim = 2000,
  seed = 2022,
  return_draws = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{d}{The data frame after transformation}

\item{all_model_coefs}{An list of all models}

\item{N_sim}{Number of simulations}

\item{seed}{Seed value for reproducibility; use default if not provided}

\item{return_draws}{logical value: if draw level prediction is returned as a matrix}
}
\value{
if return_draws is TRUE, a list including a vector of predicted probabilities
and a matrix of predictions at draw level; if FALSE, only a vector
}
\description{
Predict missing FPG based on HbA1c using an appropriate model
}
\examples{
predictions <- transform(data, models, N_sim = 2000)
}
