% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform data consistent with how models were fitted}
\usage{
transform_data(d)
}
\arguments{
\item{d}{The data frame with all necessary columns}
}
\value{
A data frame with columns of standard names
}
\description{
Transform data consistent with how models were fitted
}
\examples{
data_transformed <- transform(data)
}
